Implement spinning logic using 2 reel instances
`this.reels` is a container hloding 2 identical containers of symbols.
By using 2 instances we achieve the illusion of infinite spinning.

On every tween update we first calculate `totalY` - the absolute
position on the whole spinning path. Then we calculate
`instancePositions[i]` that represent a relative position
(on the circumference).

The instances y positions are then set so they always follow each other
in the order defined by the `instancePositions` values.

`DEV_MODE_getColor` helps visualise the algorithm.